Removal of AI from notepad, and paint.
I remember in your original video on Talon you talked about Microsoft adding AI features to notepad, and you seemed to dislike the idea of that, so I think it would be a good addition to disable those features from both notepad and paint. I also added in the disabling of 365 ads because I don't think anybody is a fan of ads.